got the card swiper (msr) working.  I gave in and used a windows form to read the card input.  I also added a isCcIntegrationOn boolean.  If this is false, the software will add credit card payments without making the user trap the input.  I still plan to add a preferences page and table to store some of these settings. RESOLVED BBQRMS-98

+++ trunk/BBQRMSSolution/Client/Controls/NumericTextBox.cs
@@ -13,8 +13,8 @@ namespace Controls
 
         private Boolean CheckFormat(string text)
         {
-            short val;
-            return Int16.TryParse(text, out val);
+            Decimal val;
+            return Decimal.TryParse(text, out val);
         }
 
         private void CheckPasteFormat(object sender, DataObjectPastingEventArgs e)
